A scope always has a compiler.
The scope object contained some internal code to substitute an external
environment when the compiler was not supplied.  This was used only in
testing, not in any production capacity.

In light of that, we can eliminate the dynamic decision making inside the
scope and simply demand that a compiler instance (or a fake equivalent) is
always supplied.

This reduces the complexity of the code and makes clearer the object
relationships involved.

(The small cost is a lot of testing that depended on this had to change to use
 a real compiler.  Most of the change derives from that.)

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>