Use `Forwardable` rather than hand-rolling it.
Using the standard `forwardable` library makes for a clearer understand of
what is going on where in our code - if we just wrap directly around a method
on a member, we might as well delegate properly.

That way people viewing the code know, with certainty, that it isn't actually
doing anything of significance other than passing that call directly on.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>