Use `alias` to map []{,=} to {lookup,set}var on scope.
Rather than hand-rolling an alias for the hash-like accessors to the named get
and set methods, we can just use the stock alias method.

This eliminates an extra layer of function calls, but also makes clearer to
code analysis and the casual reader what the actual significance of those
different methods are.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>