Treat `nil` as false, not true in scope and AST evaluation.
We had a central place defined in the scope to determine the truthiness of
objects from an AST / code evaluation point of view.  Since inception this
treated a `nil` value as *true*, in variance to Ruby behaviour.

While odd, this probably doesn't make sense to preserve, so add a test to
validate that `nil` becomes false, and stays that way.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>