Use `set_options` rather than open-coding it everywhere.
`Puppet::Util::MethodHelper` has a handful of tools for handling the common
pattern of accepting a hash and setting object attributes from it.  We also
had a bunch of places in the code that open-coded this same pattern.

This just replaces them all with the single, standard implementation.

Really, we should probably even hoist this up to something like Object or
Module so that it is available everywhere to peer with the `attr_*` methods
and all.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>