ast: eliminate dead `settor` attribute of AST nodes.
Once upon a time, AST entries that set something were evaluated before entries
that didn't.  That was eliminated on Fex 28, 2008, in fd1573fd, by Luke.

We might as well take the last, and misleading, vestige of that out of the
tree at this point.

Also, eliminate a couple of comment lines that have lingered from that era
attached to something randomly else.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>