lexer: be consistent about how to access a member
If you are going to use the accessor method rather than the member variable to
check if something is set, you should probably also use it to determine what
to return, right?

Meanwhile, collapse the entire statement into a single `a or b` expression,
because everyone knows there is an international shortage of newlines.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>