lexer: use methodhelper to set token options.
The code had an open-coded version of the set_options code from MethodHelper,
with the trivial variance that the setter was invoked externally rather
than internally.

That doesn't have much point - might as well pass the options to the token and
have it internally handle setting all the values instead.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>