(#15533) Change puppet module build to build the module in cwd (or parents)
Previously 'puppet module build' required a full path to build a module. This
change will add the ability to run the command from within a module itself
and have the package built correctly.

Through changes to find_module_root it will also support running the command
is a sub-directory of the module as well.

Due to some lack of system and unit testing, I've added that as well as part
of this patch and included tests for the new behaviour.