scope: simplify and clarify if statements
Reading code that starts with a negation - the true case flows through to the
else block - is harder than reading normal, linear code.

This reverses an unless to if to reduce the cognitive load of reading the code.

It also rewrites a `return foo if bar; baz` into a simple if for the same
clarity reasons.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>