Eliminate Puppet type attribute aliases.
About six years back we added attribute aliases to the Puppet type definition.
Although it isn't completely clear why this was done, it wasn't used then, and
hasn't been used over the last six years or so.

Since this has a non-trivial cost checking for an alias every single time we
touch a property, and we do that ... a lot, eliminating the code has a small
but notable reduction in CPU time.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>