Disable ActiveSupport::Dependencies hooks.
The ActiveSupport::Dependencies code is loaded along with Rails, and hooks
into a bunch of the system.  This consumes a surprisingly large amount of CPU
time watching for constants and generally acting to support unloading code.

This is related to the code that automatically requires libraries when the
constant is first accessed.

We don't - can't, since Rails is a soft dependencies - depend on that feature,
or anything related to it, we have no benefit from paying that horrible price.

Dropping this is actually a significant drop in CPU consumption - some 8
percent of the CPU cost in my testing is attributed directly to this feature.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>