(Maint) Don't assume paths are absolute
Previously, the spec test failed on Windows, because the test assumed that
paths such as '/sbin/rc-service' are absolute on every platform. Since
they are not absolute on Windows, the test would fail.

This commit changes the test to not rely on the internals of the
`Puppet::Util.which` command that is included in every provider.