cr 15547
Unless we are able to mock the operating system Fact, we will not be
able to run all the mount parsing tests running on each platforms.

However this is not required. We could aswell run the platform specific
test on the specific platform. However this is not done at present. We
are just making it pending for solaris since other platforms pass.
(TODO)

for the mount_spec, the mocked operating system value is not reflected
in lib/puppet/provider/mount/parsed.rb needing @blockdevice to be passed
in even though we are impersonating as Darwin

for the posix_spec, solaris requries valid values for locale
environments. So we change es_ES to en_US as the sentinel value.