Eliminate remaining uses of FileCollection.
As noted, the whole FileCollection concept was an overall loss - it used more
memory, more CPU, and added more complexity.

This eliminates the rest of the uses of the tool, and turns it into an empty
stub that avoids breaking the external API of things that included it, but
removes any unification that it tried to do.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>