Retrieve files from Fileserver with static compiler
This patchs ables puppet agent to retrieve file content either from filebucket
if a checksum is present, or from fileserver if a source definition is present.
Both could exist in catalog if static compiler is enabled. In that case,
priority between this two modes is controled using the new option
content_origin_priority.