Revert "Merge pull request #920 from rtyler/issues/gemfile-15464"
This reverts commit edc36fc7351630158828305e5124e5b46f89cbdf, reversing
changes made to 849f31a605a886b2246c09b7c008835e16010b78.

The gemfile is a great idea, but the one that was merged in is
incorrect. I shows a dependency on cucumber for development and another
dependency on rack. If we are going to bring this in it should reflect
our actual dependencies and we need to get it used in our testing
environments so that it stays up to date.