8200 : validate only if we can (especially path)
The global validate block is called even when `puppet resource zone` is
called without any further arguments, and there is no good way to detect
when that happens. More over, the `isrequired` method in parameter.rb is
a better way to mark the parameter :path as a required option. Hence the
validation of path is removed from global validate block.

The ip validation is refactored to make it clearer to read.

Since `isrequired` doesnot seem to be hooked up, the spec test for path
validation is now in pending.