Remove puppetlabs_spec_helper require from root Rakefile
Having puppetlabs_spec_helper as a require in the main Rakefile means that any
rake tasks require the library, which is undesirable as it pulls in many other
dependencies, including mocha, rspec-puppet, etc. Needing these gems installed
to release a gem or build a release package is undesireable as well, so this
commit removes that require. There is still a require on puppetlabs_spec_helper
in spec/spec_helper.rb.