Re-introduce a Gemfile for managing dependencies for puppet contributors
The Gemfile is only for use by contributors to the project, it has no effect on
the packaging of Puppet as a gem or anything else. It's sole purpose is to make
it easier for a new developer to get all the dependencies needed to develop for Puppet installed locally.

Pairs quite nicely when used with RVM: <http://rvm.io>

Fixes #15464