(Maint) Remove some more ambiguity
Remove the pieces of the document that say we support email and
Redmine patches, instead of just saying that we strongly
discourage these. Also change the places where we say "we recommend"
when what we really mean is "we require". Lastly, make a few small
changes that were brought up on the PR.

<hailee@puppetlabs.com>