(#15676) Fix puppet module face SSL acceptance tests to use masters real hostname
Previously we had been using 'localhost' as the location of the master, which
is a dangerous assumption with systest, since the master may be located on
another host or 'localhost' may not give us the expected results.

This patch changes the usage of localhost to use the variable 'master' which
should point to the canonical hostname of the master each and every time.