(#15668) Replace test for order of files in pluginsync
The previous version of this test relied on using puppet apply to check
that the module path order was preserved in deciding the files to use
from a pluginsync. Without apply performaning a pluginsync that was not
possible anymore, but this test is still useful.

This replaces the previous test with a new one that uses the master and
the agent for performing a pluginsync and checking what files the agent
recieves.