(#11870) Refactor the MSI provider to use package options
Previously, the MSI provider contained logic for accepting a hash of
`install_options`. The provider's install method flattened the hash pairs
as:

    "key1=value1 key2=value2"

and passed that to msiexec.exe. This worked fine for MSI public
properties, but does not account for passing single arguments.

This commit refactors the provider to use the more generic version of
`install_options`, which expects an array of values. For example, if a
manifest contains:

    [ '/S', { 'INSTALLDIR' => 'C:\Program Files' } ]

then msiexec will be invoked with:

    "/S INSTALLDIR=\"C:\\Program Files\""

The only change in behavior is that the MSI provider will now accept
single arguments and pass those onto msiexec.