(#11870) Add the ability to specify uninstall_options
Previously, the MSI provider was unable to specify uninstall_options.
This isn't typically an issue, since Windows stores the command and
arguments necessary to uninstall the package in the `UninstallString`
registry value associated with the package.

But there are some MSI packages that require this functionality, e.g. to
specify VMware Tools modules to uninstall:

   msiexec.exe /x <productcode> REMOVE=Sync,VSS,...

This commit adds an `uninstall_options` parameter to the package type and
updates the MSI provider to use it. It has the same format as
`install_options`.