(#11870) Deprecate the `:msi` provider
The `:msi` package provider is being deprecated and replaced with the
`:windows` provider. Attempting to install or uninstall a package using
the msi provider will cause puppet to issue a deprecation warning.

The windows provider is basically a copy of the MSI provider, though it
will be capable of managing executable packages in a future commit.