(#11870) Refactor package management logic for MSI
This commit moves the logic for querying, installing and uninstalling MSI
packages into a MsiPackage class separate from the main windows package
provider. This is being done in preparation for adding executable package
support (as some of the logic is very similar), such as querying the
registry for installed products. Additionally, executable and MSI
packages can be related, which makes partitioning the set of installed
packages into different MSI and executable providers not really feasible.