Add HOWTO hint on topic branches in CONTRIBUTING
Without this patch it's not very clear how to create a topic branch.
This has been a problem in the past because we often get pull requests
from the master branch to the master branch which can be confusing when
working outside of the GitHub Web UI.

This patch addresses the problem by providing a concrete example of how
to create a reasonable topic branch and start using it with two
commands.