(#7962) Added settings class for handling time intervals.
The DurationSetting converts strings with units into a number of
seconds, but more importantly guarantees that the setting will be an
integer. The following settings can now be specified as durations:

  configtimeout
  filetimeout
  waitforcert
  splaylimit
  runinterval
  rrdinterval
  ca_ttl

Since multiple settings can use this format, the description was removed
from `ca_ttl` to be added to the config section of puppet-docs.

Also note, since many of these settings specify sub-hour times, the unit
'm' was added for specifying times in minutes.