(#10915) Fix tests to work with new parameter checking
The change to how the property checks if the provider can be called
caused several tests to fail because they were not using providers that
were able to properly report `respond_to?`. This fixes those tests. In
the case of the host_spec.rb I was able to convert away from a mock, but
in the mount_spec.rb the mock had to mock `respond_to?` since the
interaction with the provider seems to be more complicated than I can
unravel.