(#9416) Confine based on the presence of a `zpool` command
Previously, the zpool provider was confined to solaris and assumed the
existence of `/usr/sbin/zpool`. This commit changes the command to a
relative path, which will use `Puppet::Util.which` to resolve the
command. The default search path is set by the `path` setting in
puppet.conf.

Based on work by Marc Fournier <marc.fournier@camptocamp.com>