(#15644) refactor the parse_line so that it supports the newer format
Previously the parse_line conflated the different output formats of the
older and newer `pkg list` commands. This commit splits the formats to
separate regexes so that it is easier to manage. Especially if the pkg
list format changes yet again.

This commit simplifies self.instances, latest and query methods by
removing redundant lines.

Adds more comments to the methods, and changes the method names to reflect the
method usage.

Tests:
- Update the package list output so that the format is correct
(it is ufoxi - 5 dashes)

- Earlier, when a pkg list line could not be parsed, the provider warned
the user about this fact and continued. This behavior is dangerous. We
change it to fail instead.

- Added more test cases.