(#15686) Ensure exception message is a string
Previously the execpipe method used to return an array as output, which
if the child exit status was non-zero, caused an exception to be raised
with the array as the error message, instead of a string.

Also, previously, we relied on exception handling to know when a package
is not present. But there are cases where no exception is raised, but
the package output has an 'ERROR' string.

This commit changes from `execpipe` to `execute`, which will raise an
exception with the output of the `pkginfo` command.  It also explicitly
looks for an 'ERROR' string in the package output.

This commit also refactors the duplicate namemap so that it is defined
only once, and removes the nil entries so that only keys of interest are
retained.

It also adds fixtures and spec tests for the provider.