(#8200) ensure zonecfg errors are caught
zonecfg can return 0 eventhough it failed to apply a configruation. We
can catch it only if we check for the output. We now check for
/not allowed/ to verify that the property was applied.

previously, we were using using `zonecfg -z zone -f -` to read commands
from stdinput and using IO.popen. However, we do not need to do this as
we pass confing commands one at a time, and can invoke zonecfg with the
command directly. Hence removed the popen invocation from code for a
simpler exec.