Always set log functions on root environment
Set the logging parser functions (notice, err, etc) on the root environment.
This also requires setting the current environment to be nil in
Environment.clear by setting it directly on Thread.current.  Otherwise the
default environment from puppet.conf is used. In tests this was interfering
with the log functions being set on the root environment instead of the
current environment.