(#15853) Update gem dependencies to reflect rc ordering
Facter 2.0.0rc is less than 2.0.0, so >= 2.0.0 will never be satisfied using
the current Puppet gem dependencies. This commit updates those dependencies to
have rc in the version, which will be correctly handled by modern rubygems and
ignored by older rubygems. The rc can be removed once the releases are final
and we no longer want users to be depending on rc releases.